Converted useWindowWidth hook to ContextAPI
