+++ components/common/Container.js
@@ -1,8 +1,8 @@
-import React from 'react';
-import useWindowWidth from '../hooks/useWindowWidth';
+import React, { useContext } from 'react';
+import { WindowContext } from '../contexts/WindowContext';
 
 export default function Container({ children }) {
-  const { isSmallerDevice } = useWindowWidth();
+  const { isSmallerDevice } = useContext(WindowContext);
   return (
     <div
       style={{
